Corrected BirdCall and Anchor.. args types in BirdCall.scala

+++ src/main/scala/org/laem/weeki/BirdCall.scala
@@ -12,7 +12,7 @@ object BirdCall {
    * 
    * 
    */
-	val res = AnchorSearch.search(args(0), args(1).toInt)
+	val res = AnchorSearch.search(args(0), args(1))
 	println(res)
     SearchAPIClient.go(res.map(_.productElement(0).toString))
